# packages/control-sdk/openapi/control-plane.v2.json
{
  "openapi": "3.1.0",
  "info": { "title": "Agent HQ Control Plane API", "version": "2.0.0" },
  "paths": {
...
    "/v1/marketplace/catalog": {
      "post": {
        "operationId": "marketplace.catalog.read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "caller": {
                    "properties": {
                      "servicePrincipalId": {
                        "maxLength": 64,
                        "minLength": 5,
                        "pattern": "^svc_[a-z][a-z0-9-]*$",
                        "type": "string"
                      }
                    },
                    "required": ["servicePrincipalId"],
                    "type": "object"
                  },
                  "contractVersion": {
                    "additionalProperties": false,
                    "properties": {
                      "major": {
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991,
                        "type": "integer"
                      },
                      "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                    },
                    "required": ["major", "minor"],
                    "type": "object"
                  },
                  "correlation": {
                    "properties": {
                      "causationCommandId": {
                        "pattern": "^cmd_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "parentEventId": {
                        "pattern": "^evt_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "traceId": { "pattern": "^trc_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                    },
                    "required": ["traceId"],
                    "type": "object"
                  },
                  "operation": { "const": "marketplace.catalog.read", "type": "string" },
                  "parameters": {
                    "properties": {
                      "workspaceIdentity": {
                        "properties": {
                          "userId": { "maxLength": 128, "minLength": 1, "type": "string" },
                          "workspaceId": { "maxLength": 128, "minLength": 1, "type": "string" }
                        },
                        "required": ["workspaceId", "userId"],
                        "type": "object"
                      }
                    },
                    "required": ["workspaceIdentity"],
                    "type": "object"
                  },
                  "projectId": { "pattern": "^prj_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "requestedAt": {
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                    "type": "string"
                  },
                  "requestId": { "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "workspaceId": { "pattern": "^wsp_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                },
                "required": [
                  "caller",
                  "contractVersion",
                  "requestId",
                  "workspaceId",
                  "correlation",
                  "operation",
                  "requestedAt",
                  "parameters"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "artifacts": {
                          "additionalProperties": false,
                          "properties": {
                            "catalog-latest.v1.json": { "type": "string" },
                            "catalog-summary.v1.json": { "type": "string" },
                            "catalog.v1.json": { "type": "string" },
                            "categories.v1.json": { "type": "string" },
                            "compatibility.v1.json": { "type": "string" },
                            "integrity.json": { "type": "string" },
                            "sources.lock.json": { "type": "string" }
                          },
                          "required": [
                            "catalog.v1.json",
                            "catalog-latest.v1.json",
                            "catalog-summary.v1.json",
                            "categories.v1.json",
                            "compatibility.v1.json",
                            "integrity.json",
                            "sources.lock.json"
                          ],
                          "type": "object"
                        },
                        "catalogId": { "pattern": "^catalog:[a-f0-9]{64}$", "type": "string" },
                        "installations": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "canonicalContentDigest": {
                                "pattern": "^sha256:[a-f0-9]{64}$",
                                "type": "string"
                              },
                              "pluginId": {
                                "pattern": "^plugin:[a-z0-9-]+:[a-z0-9][a-z0-9-]{1,127}$",
                                "type": "string"
                              },
                              "releaseId": {
                                "pattern": "^release:[a-f0-9]{64}$",
                                "type": "string"
                              },
                              "state": {
                                "enum": [
                                  "pending-authorization",
                                  "unavailable",
                                  "rejected-by-policy",
                                  "installed",
                                  "superseded"
                                ],
                                "type": "string"
                              }
                            },
                            "required": [
                              "pluginId",
                              "releaseId",
                              "canonicalContentDigest",
                              "state"
                            ],
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "releaseId": { "pattern": "^catalog:[a-f0-9]{64}$", "type": "string" },
                        "state": { "enum": ["ready", "stale"], "type": "string" }
                      },
                      "required": ["catalogId", "releaseId", "state", "artifacts", "installations"],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "data"],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "class": {
                          "enum": [
                            "validation",
                            "authentication",
                            "authorization",
                            "conflict",
                            "stale_reference",
                            "capability_mismatch",
                            "runtime_unavailable",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "code": {
                          "maxLength": 96,
                          "minLength": 1,
                          "pattern": "^[A-Z][A-Z0-9_]*$",
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {},
                          "propertyNames": { "maxLength": 64, "type": "string" },
                          "type": "object"
                        },
                        "message": { "maxLength": 512, "minLength": 1, "type": "string" },
                        "remediation": { "maxLength": 256, "minLength": 1, "type": "string" },
                        "retryable": { "type": "boolean" },
                        "source": {
                          "enum": [
                            "client",
                            "auth",
                            "policy",
                            "runtime",
                            "provider",
                            "persistence",
                            "workflow",
                            "system"
                          ],
                          "type": "string"
                        }
                      },
                      "required": ["class", "code", "message", "retryable", "source"],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "error"],
                  "type": "object"
                }
              }
            },
            "description": "Normalized error response"
          }
        },
        "security": [{ "serviceBearer": [] }]
      }
    },
    "/v1/marketplace/install": {
      "post": {
        "operationId": "marketplace.install.request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "caller": {
                    "properties": {
                      "servicePrincipalId": {
                        "maxLength": 64,
                        "minLength": 5,
                        "pattern": "^svc_[a-z][a-z0-9-]*$",
                        "type": "string"
                      }
                    },
                    "required": ["servicePrincipalId"],
                    "type": "object"
                  },
                  "commandId": { "pattern": "^cmd_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "contractVersion": {
                    "additionalProperties": false,
                    "properties": {
                      "major": {
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991,
                        "type": "integer"
                      },
                      "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                    },
                    "required": ["major", "minor"],
                    "type": "object"
                  },
                  "correlation": {
                    "properties": {
                      "causationCommandId": {
                        "pattern": "^cmd_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "parentEventId": {
                        "pattern": "^evt_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "traceId": { "pattern": "^trc_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                    },
                    "required": ["traceId"],
                    "type": "object"
                  },
                  "idempotencyKey": {
                    "maxLength": 128,
                    "minLength": 16,
                    "pattern": "^[A-Za-z0-9._:-]+$",
                    "type": "string"
                  },
                  "issuedAt": {
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                    "type": "string"
                  },
                  "operation": { "const": "marketplace.install.request", "type": "string" },
                  "payload": {
                    "additionalProperties": false,
                    "properties": {
                      "canonicalContentDigest": {
                        "pattern": "^sha256:[a-f0-9]{64}$",
                        "type": "string"
                      },
                      "pluginId": {
                        "pattern": "^plugin:[a-z0-9-]+:[a-z0-9][a-z0-9-]{1,127}$",
                        "type": "string"
                      },
                      "releaseId": { "pattern": "^release:[a-f0-9]{64}$", "type": "string" },
                      "requestedHarness": { "maxLength": 128, "minLength": 1, "type": "string" },
                      "workspaceIdentity": {
                        "properties": {
                          "userId": { "maxLength": 128, "minLength": 1, "type": "string" },
                          "workspaceId": { "maxLength": 128, "minLength": 1, "type": "string" }
                        },
                        "required": ["workspaceId", "userId"],
                        "type": "object"
                      }
                    },
                    "required": [
                      "pluginId",
                      "releaseId",
                      "canonicalContentDigest",
                      "requestedHarness",
                      "workspaceIdentity"
                    ],
                    "type": "object"
                  },
                  "payloadHash": { "pattern": "^[a-f0-9]{64}$", "type": "string" },
                  "projectId": { "pattern": "^prj_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "requestId": { "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "workspaceId": { "pattern": "^wsp_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                },
                "required": [
                  "caller",
                  "contractVersion",
                  "requestId",
                  "workspaceId",
                  "correlation",
                  "commandId",
                  "idempotencyKey",
                  "payloadHash",
                  "operation",
                  "issuedAt",
                  "payload"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "canonicalContentDigest": {
                          "pattern": "^sha256:[a-f0-9]{64}$",
                          "type": "string"
                        },
                        "catalogId": { "pattern": "^catalog:[a-f0-9]{64}$", "type": "string" },
                        "installationId": { "maxLength": 128, "minLength": 1, "type": "string" },
                        "pluginId": {
                          "pattern": "^plugin:[a-z0-9-]+:[a-z0-9][a-z0-9-]{1,127}$",
                          "type": "string"
                        },
                        "releaseId": { "pattern": "^release:[a-f0-9]{64}$", "type": "string" },
                        "requestedHarness": { "maxLength": 128, "minLength": 1, "type": "string" },
                        "state": {
                          "enum": [
                            "pending-authorization",
                            "unavailable",
                            "rejected-by-policy",
                            "installed",
                            "superseded"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "pluginId",
                        "releaseId",
                        "canonicalContentDigest",
                        "installationId",
                        "catalogId",
                        "requestedHarness",
                        "state"
                      ],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "data"],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "class": {
                          "enum": [
                            "validation",
                            "authentication",
                            "authorization",
                            "conflict",
                            "stale_reference",
                            "capability_mismatch",
                            "runtime_unavailable",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "code": {
                          "maxLength": 96,
                          "minLength": 1,
                          "pattern": "^[A-Z][A-Z0-9_]*$",
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {},
                          "propertyNames": { "maxLength": 64, "type": "string" },
                          "type": "object"
                        },
                        "message": { "maxLength": 512, "minLength": 1, "type": "string" },
                        "remediation": { "maxLength": 256, "minLength": 1, "type": "string" },
                        "retryable": { "type": "boolean" },
                        "source": {
                          "enum": [
                            "client",
                            "auth",
                            "policy",
                            "runtime",
                            "provider",
                            "persistence",
                            "workflow",
                            "system"
                          ],
                          "type": "string"
                        }
                      },
                      "required": ["class", "code", "message", "retryable", "source"],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "error"],
                  "type": "object"
                }
              }
            },
            "description": "Normalized error response"
          }
        },
        "security": [{ "serviceBearer": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "service-credential" }
    }
  }
}